feat: auto install missing dependencies
